requirements.txt
matplotlib==3.3.4
numpy==1.23.0
opencv-python==4.8.0.74
Pillow==8.4.0
tqdm==4.64.1
wandb==0.13.5
torch
torchvision
tensorboard
future
openpyxl
fiona
rasterio